+++ packages/views/onboarding/index.ts
@@ -1,4 +1,3 @@
 export { OnboardingFlow, type OnboardingStep } from "./onboarding-flow";
 export { CliInstallInstructions } from "./steps/cli-install-instructions";
-export { StarterContentPrompt } from "./components/starter-content-prompt";
 export { CloudWaitlistExpand } from "./components/cloud-waitlist-expand";
